--- faces/float.go
@@ -1,6 +1,0 @@
-package faces
-
-// Float ...
-type Float interface {
-	Float() float64
-}
