--- faces/gt.go
@@ -1,6 +1,0 @@
-package faces
-
-// GreaterThan ...
-type GreaterThan interface {
-	GreaterThan(b interface{}) (bool, error)
-}
