--- faces/gte.go
@@ -1,6 +1,0 @@
-package faces
-
-// GreaterThanEqualTo ...
-type GreaterThanEqualTo interface {
-	GreaterThanEqualTo(interface{}) (bool, error)
-}
