--- faces/int.go
@@ -1,6 +1,0 @@
-package faces
-
-// Int ...
-type Int interface {
-	Int() int
-}
