--- faces/lt.go
@@ -1,6 +1,0 @@
-package faces
-
-// LessThan ...
-type LessThan interface {
-	LessThan(b interface{}) (bool, error)
-}
