--- faces/lte.go
@@ -1,6 +1,0 @@
-package faces
-
-// LessThanEqualTo ...
-type LessThanEqualTo interface {
-	LessThanEqualTo(interface{}) (bool, error)
-}
