--- faces/map.go
@@ -1,6 +1,0 @@
-package faces
-
-// Map ...
-type Map interface {
-	Map() map[interface{}]interface{}
-}
