--- faces/ne.go
@@ -1,7 +1,0 @@
-package faces
-
-// NotEqual defines an interface to support the
-// checking "equality" of one type to an another.
-type NotEqual interface {
-	NotEqual(b interface{}) (bool, error)
-}
