--- faces/slice.go
@@ -1,6 +1,0 @@
-package faces
-
-// Slice ...
-type Slice interface {
-	Slice() []interface{}
-}
