--- faces/value.go
@@ -1,6 +1,0 @@
-package faces
-
-// Value ...
-type Value interface {
-	Value() string
-}
