--- types/value.go
@@ -1,25 +1,0 @@
-package types
-
-import (
-	"fmt"
-
-	"github.com/gobuffalo/lush/faces"
-)
-
-// Valuer ...
-type Valuer string
-
-// Value ...
-func (s Valuer) Value() string {
-	return fmt.Sprintf("%v", s)
-}
-
-// Value ...
-func Value(i interface{}) string {
-	switch t := i.(type) {
-	case faces.Value:
-		return t.Value()
-	default:
-		return fmt.Sprintf("%v", i)
-	}
-}
